// g4s-spa-module/src/app/modules/request/models/creating-connection-request.model.ts
export class CreatingConnectionRequest {
    player: string;
    target: string;
    playerToTargetMessage: string;
    currentStatus: string;
    strength: number;    
    tags: string[];
    middleMan: string;
    playerToMiddleManMessage: string;
    middleManToTargetMessage: string;
}